chore: Update docker-compose.yml to add frpc_azure service
This commit adds the `frpc_azure` service to the `docker-compose.yml` file. The service is configured with the `localhost:8080/box_starting/local_frp` image and is connected to the `frp_net` network. It also sets the `FRP_SERVER` environment variable to `20.194.196.94` and mounts the `frp-token` secret. This service should not be auto redeployed and may require manual restart during incidents.