chore: Add webdav service to docker-compose.yml
This commit adds the `webdav` service to the `docker-compose.yml` file. The service is configured with the `hub.aiursoft.cn/aiursoft/static` image and is connected to the `proxy_app` network. It also mounts the `media-data` volume and sets the entrypoint to run the `/app/static` command with specific parameters.